RawCord/requirements.txt
requests==2.26.0
websocket_client==1.2.1
